requirements/prod.txt
-r common.txt

gunicorn
psycopg2-binary
django-storages
boto3
mysqlclient